signUp fix

+++ src/components/form/formTechsUpdate/index.js
@@ -92,47 +92,51 @@ const FormTechsUpdate = () => {
 
   const handleRemoveTech = (e) => {
     e.preventDefault();
-
-    const actualTech = user.techs.filter((actual) => actual.title === tech);
-    const id = actualTech[0].id;
-    const headers = {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    };
-    axios.delete(`${baseUrl}users/techs/${id}`, headers).then((res) => {
-      axios.get(`${baseUrl}users/${user.id}`).then((res) => {
-        dispatch(loginThunk(res.data));
+    if (tech) {
+      const actualTech = user.techs.filter((actual) => actual.title === tech);
+      const id = actualTech[0].id;
+      const headers = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
+      axios.delete(`${baseUrl}users/techs/${id}`, headers).then((res) => {
+        axios.get(`${baseUrl}users/${user.id}`).then((res) => {
+          dispatch(loginThunk(res.data));
+        });
       });
-    });
+    }
   };
 
   const handleUpdateTech = (e) => {
     e.preventDefault();
+    if (attTech) {
+      const data = {
+        title: attTech,
+        status: attTechStatus,
+      };
 
-    const data = {
-      title: attTech,
-      status: attTechStatus,
-    };
-
-    const actualTech = user.techs.filter((actual) => actual.title === attTech);
-    console.log(actualTech);
-    const id = actualTech[0].id;
-    const headers = {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    };
-    axios.put(`${baseUrl}users/techs/${id}`, data, headers).then((res) => {
-      axios
-        .get(`${baseUrl}users/${user.id}`)
-        .then((res) => {
-          dispatch(loginThunk(res.data));
-        })
-        .then((res) => {
-          window.location.reload();
-        });
-    });
+      const actualTech = user.techs.filter(
+        (actual) => actual.title === attTech
+      );
+      console.log(actualTech);
+      const id = actualTech[0].id;
+      const headers = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
+      axios.put(`${baseUrl}users/techs/${id}`, data, headers).then((res) => {
+        axios
+          .get(`${baseUrl}users/${user.id}`)
+          .then((res) => {
+            dispatch(loginThunk(res.data));
+          })
+          .then((res) => {
+            window.location.reload();
+          });
+      });
+    }
   };
 
   return (
@@ -184,7 +188,6 @@ const FormTechsUpdate = () => {
           Enviar
         </Button>
       </form>
-
       {user.techs.length !== 0 && (
         <>
           <form onSubmit={handleUpdateTech} className={classes.form}>
