+++ src/components/form/formWorksUpdate/index.js
@@ -83,42 +83,46 @@ const FormWorksUpdate = () => {
 
   const handleRemoveWork = (e) => {
     e.preventDefault();
-
-    const actualWork = user.works.filter((actual) => actual.title === work);
-    const id = actualWork[0].id;
-    const headers = {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    };
-    axios.delete(`${baseUrl}users/works/${id}`, headers).then((res) => {
-      axios.get(`${baseUrl}users/${user.id}`).then((res) => {
-        dispatch(loginThunk(res.data));
+    if (work) {
+      const actualWork = user.works.filter((actual) => actual.title === work);
+      const id = actualWork[0].id;
+      const headers = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
+      axios.delete(`${baseUrl}users/works/${id}`, headers).then((res) => {
+        axios.get(`${baseUrl}users/${user.id}`).then((res) => {
+          dispatch(loginThunk(res.data));
+        });
       });
-    });
+    }
   };
 
   const handleUpdateWork = (e) => {
     e.preventDefault();
+    if (attWork) {
+      const data = {
+        title: workTitle,
+        description: workDescription,
+        deploy_url: workUrl,
+      };
 
-    const data = {
-      title: workTitle,
-      description: workDescription,
-      deploy_url: workUrl,
-    };
-
-    const actualWork = user.works.filter((actual) => actual.title === attWork);
-    const id = actualWork[0].id;
-    const headers = {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    };
-    axios.put(`${baseUrl}users/works/${id}`, data, headers).then((res) => {
-      axios.get(`${baseUrl}users/${user.id}`).then((res) => {
-        dispatch(loginThunk(res.data));
+      const actualWork = user.works.filter(
+        (actual) => actual.title === attWork
+      );
+      const id = actualWork[0].id;
+      const headers = {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      };
+      axios.put(`${baseUrl}users/works/${id}`, data, headers).then((res) => {
+        axios.get(`${baseUrl}users/${user.id}`).then((res) => {
+          dispatch(loginThunk(res.data));
+        });
       });
-    });
+    }
   };
 
   return (
